lecture-09/exercise_DL_pcg/requirements.txt
certifi==2021.5.30
click==8.0.1
cycler==0.10.0
importlib-metadata==4.8.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
torch==1.9.0
tqdm==4.62.2
typing-extensions==3.10.0.2
zipp==3.5.0
